refactor: Avoid unnecessary call to setRequestHeaders after matching
