+++ lib/playback_interceptor.js
@@ -8,46 +8,6 @@ const _ = require('lodash')
 const common = require('./common')
 const DelayedBody = require('./delayed_body')
 
-function setHeader(request, name, value) {
-  debug('setHeader', name, value)
-
-  request.setHeader(name.toLowerCase(), value)
-
-  // https://github.com/nock/nock/issues/256
-  if (name === 'expect' && value === '100-continue') {
-    timers.setImmediate(function() {
-      debug('continue')
-      request.emit('continue')
-    })
-  }
-}
-
-//  Sets request headers of the given request. This is needed during both matching phase
-//  (in case header filters were specified) and mocking phase (to correctly pass mocked
-//  request headers).
-function setRequestHeaders(req, options, interceptor) {
-  //  If a filtered scope is being used we have to use scope's host
-  //  in the header, otherwise 'host' header won't match.
-  //  NOTE: We use lower-case header field names throughout Nock.
-  const HOST_HEADER = 'host'
-  if (interceptor.__nock_filteredScope && interceptor.__nock_scopeHost) {
-    options.headers[HOST_HEADER] = interceptor.__nock_scopeHost
-    setHeader(req, HOST_HEADER, interceptor.__nock_scopeHost)
-  } else {
-    //  For all other cases, we always add host header equal to the
-    //  requested host unless it was already defined.
-    if (options.host && !req.getHeader(HOST_HEADER)) {
-      let hostHeader = options.host
-
-      if (options.port === 80 || options.port === 443) {
-        hostHeader = hostHeader.split(':')[0]
-      }
-
-      setHeader(req, HOST_HEADER, hostHeader)
-    }
-  }
-}
-
 function parseJSONRequestBody(req, requestBody) {
   if (!requestBody || !common.isJSONContent(req.headers)) {
     return requestBody
@@ -132,8 +92,6 @@ function playbackInterceptor({
   }
 
   function start() {
-    // We again set request headers, now for our matched interceptor.
-    setRequestHeaders(req, options, interceptor)
     interceptor.req = req
     req.headers = req.getHeaders()
 
@@ -389,8 +347,4 @@ function playbackInterceptor({
   start()
 }
 
-module.exports = {
-  setHeader,
-  setRequestHeaders,
-  playbackInterceptor,
-}
+module.exports = { playbackInterceptor }
