+++ lib/request_overrider.js
@@ -13,11 +13,22 @@ const timers = require('timers')
 const common = require('./common')
 const globalEmitter = require('./global_emitter')
 const Socket = require('./socket')
-const {
-  setHeader,
-  setRequestHeaders,
-  playbackInterceptor,
-} = require('./playback_interceptor')
+const { playbackInterceptor } = require('./playback_interceptor')
+
+function setHeader(request, name, value) {
+  debug('setHeader', name, value)
+
+  request.setHeader(name.toLowerCase(), value)
+
+  // https://github.com/nock/nock/issues/256
+  // Is there a gurantee this runs only once per request?
+  if (name === 'expect' && value === '100-continue') {
+    timers.setImmediate(() => {
+      debug('continue')
+      request.emit('continue')
+    })
+  }
+}
 
 /**
  * Given a group of interceptors, appropriately route an outgoing request.
@@ -191,6 +202,38 @@ class InterceptedRequestRouter {
     this.emitError(connResetError)
   }
 
+  /**
+   * Set request headers of the given request. This is needed both during the
+   * routing phase, in case header filters were specified, and during the
+   * interceptor-playback phase, to correctly pass mocked request headers.
+   * However, setting during routing means they will remain set during
+   * interceptor playback.
+   */
+  setHostHeaderUsingInterceptor(interceptor) {
+    const { req, options } = this
+
+    // If a filtered scope is being used we have to use scope's host in the
+    // header, otherwise 'host' header won't match.
+    // NOTE: We use lower-case header field names throughout Nock.
+    const HOST_HEADER = 'host'
+    if (interceptor.__nock_filteredScope && interceptor.__nock_scopeHost) {
+      options.headers[HOST_HEADER] = interceptor.__nock_scopeHost
+      setHeader(req, HOST_HEADER, interceptor.__nock_scopeHost)
+    } else {
+      // For all other cases, we always add host header equal to the requested
+      // host unless it was already defined.
+      if (options.host && !req.getHeader(HOST_HEADER)) {
+        let hostHeader = options.host
+
+        if (options.port === 80 || options.port === 443) {
+          hostHeader = hostHeader.split(':')[0]
+        }
+
+        setHeader(req, HOST_HEADER, hostHeader)
+      }
+    }
+  }
+
   startPlayback() {
     debug('ending')
     this.playbackStarted = true
@@ -209,9 +252,7 @@ class InterceptedRequestRouter {
     })
 
     interceptors.forEach(interceptor => {
-      // For correct matching we need to have correct request headers - if
-      // these were specified.
-      setRequestHeaders(req, options, interceptor)
+      this.setHostHeaderUsingInterceptor(interceptor)
     })
 
     const requestBodyBuffer = Buffer.concat(this.requestBodyBuffers)
